Added maze expansion and robot movements. Walls do not yet work.
The maze now expands using numpy methods when the robot attempts
to move off the display grid. This was accomplished using the
following equation to calculate display grid coordinates from
robot coordinates.

pos = (x, y)
grid_pos = offsets + (x * 2, y * -2)

The negative two was necessary in order to ensure that north was
represented as upwards in the display.